separate slice into internal and refcount, with refcount being the minimal base
